+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+- Added
+  - Additional commands for SSAs (copying account/key IDs to clipboard, and viewing key details)
+
 ## [2.10.0] - 2025-10-07
 
 - Added